fix article/button - hero

+++ src/components/Demo.jsx
@@ -3,6 +3,7 @@ import linkIcon from "../assets/link.svg";
 import { useLazyGetSummaryQuery } from "../services/article";
 import copy from "../assets/copy.svg";
 import loader from "../assets/loader.svg";
+import tick from "../assets/tick.svg";
 
 const Demo = () => {
   const [article, setArticle] = useState({
@@ -12,8 +13,10 @@ const Demo = () => {
   const [allArticles, setAllArticles] = useState([]);
   const [copied, setCopied] = useState("");
 
+  //lazy query
   const [getSummary, { error, isFetching }] = useLazyGetSummaryQuery();
 
+  // Load data from localStorage on mount
   useEffect(() => {
     const articlesFromLocalStorage = JSON.parse(
       localStorage.getItem("articles")
@@ -27,19 +30,40 @@ const Demo = () => {
   const handleSubmit = async (e) => {
     e.preventDefault();
 
-    const { data } = await getSummary({ articleUrl: article.url });
+    const existingArticle = allArticles.find(
+      (item) => item.url === article.url
+    );
 
+    if (existingArticle) return setArticle(existingArticle);
+
+    const { data } = await getSummary({ articleUrl: article.url });
     if (data?.summary) {
       const newArticle = { ...article, summary: data.summary };
       const updatedAllArticles = [newArticle, ...allArticles];
 
+      // update state and local storage
       setArticle(newArticle);
       setAllArticles(updatedAllArticles);
       localStorage.setItem("articles", JSON.stringify(updatedAllArticles));
     }
   };
+
+  // copy the url and toggle the icon
+  const handleCopy = (copyUrl) => {
+    setCopied(copyUrl);
+    navigator.clipboard.writeText(copyUrl);
+    setTimeout(() => setCopied(false), 3000);
+  };
+
+  const handleKeyDown = (e) => {
+    if (e.keyCode === 13) {
+      handleSubmit(e);
+    }
+  };
+
   return (
     <section className="mt-16 w-full max-w-xl">
+      {/* Search */}
       <div className="flex flex-col w-full gap-2">
         <form
           className="relative flex justify-center items-center"
@@ -47,39 +71,41 @@ const Demo = () => {
         >
           <img
             src={linkIcon}
-            alt="link_icon"
+            alt="link-icon"
             className="absolute left-0 my-2 ml-3 w-5"
           />
+
           <input
             type="url"
-            placeholder="Enter a URL"
+            placeholder="Paste the article link"
             value={article.url}
             onChange={(e) => setArticle({ ...article, url: e.target.value })}
+            onKeyDown={handleKeyDown}
             required
             className="url_input peer"
           />
           <button
             type="submit"
-            className="submit_btn
-          peer-focus:border-gray-700 peer-focus:text-gray-700"
+            className="submit_btn peer-focus:border-gray-700 peer-focus:text-gray-700 "
           >
-            ↵
+            <p>↵</p>
           </button>
         </form>
-        {/* History */}
-        <div className="flex flex-col gap-1m max-h-60 overflow-y-auto">
-          {allArticles.map((item, index) => (
+
+        {/* Browse History */}
+        <div className="flex flex-col gap-1 max-h-60 overflow-y-auto">
+          {allArticles.reverse().map((item, index) => (
             <div
-              key={"link-${index}"}
+              key={`link-${index}`}
               onClick={() => setArticle(item)}
               className="link_card"
             >
-              <div className="copy_btn">
+              <div className="copy_btn" onClick={() => handleCopy(item.url)}>
                 <img
-                  src={copy}
-                  alt="copy_icon"
+                  src={copied === item.url ? tick : copy}
+                  alt={copied === item.url ? "tick_icon" : "copy_icon"}
                   className="w-[40%] h-[40%] object-contain"
-                ></img>
+                />
               </div>
               <p className="flex-1 font-satoshi text-blue-700 font-medium text-sm truncate">
                 {item.url}
@@ -88,13 +114,14 @@ const Demo = () => {
           ))}
         </div>
       </div>
-      {/* Results */}
+
+      {/* Display Result */}
       <div className="my-10 max-w-full flex justify-center items-center">
         {isFetching ? (
           <img src={loader} alt="loader" className="w-20 h-20 object-contain" />
         ) : error ? (
           <p className="font-inter font-bold text-black text-center">
-            Well, unlucky.
+            Well, that wasn't supposed to happen...
             <br />
             <span className="font-satoshi font-normal text-gray-700">
               {error?.data?.error}
